refactor: move `loadConfig` into its own file to make it easier to test
